ai_microservice/requirements.txt
pydantic~=2.11.1
requests~=2.32.3
python-dotenv~=1.0.1
fastapi~=0.115.12
protobuf~=5.29.3